requirements.txt
alembic==1.10.3
anyio==3.6.2
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
click==8.1.3
cloudinary==1.32.0
decorator==5.1.1
executing==1.2.0
fastapi==0.95.1
graphql-core==3.2.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
ipython==8.12.0
jedi==0.18.2
libcst==0.4.9
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mypy-extensions==1.0.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.7
Pygments==2.15.0
python-dateutil==2.8.2
python-decouple==3.8
python-multipart==0.0.6
PyYAML==6.0
rich==13.3.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.41.0
sqlalchemy2-stubs==0.0.2a33
sqlmodel==0.0.8
stack-data==0.6.2
starlette==0.26.1
strawberry-graphql==0.171.1
traitlets==5.9.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
wcwidth==0.2.6
